validators 29/4/20 22:30

+++ comoEnCasa/comoEnCasa.COMMON/Entidades/alarmasCantidades.cs
@@ -8,7 +8,7 @@ namespace comoEnCasa.COMMON.Entidades
     {
         public int IdAlarma { get; set; }
         public int IdStockProducto { get; set; }
-        public int CantMinima { get; set; }
+        public short CantMinima { get; set; }
 
     }
 }
